[Refactor] NavigationAppearance를 BaseNavigationController에서 적용하는 것으로 수정
