+++ Favor/Favor/Sources/Flows/AuthFlow.swift
@@ -8,6 +8,7 @@
 import UIKit
 
 import FavorCoreKit
+import FavorUIKit
 import RxFlow
 
 final class AuthFlow: Flow {
@@ -16,8 +17,8 @@ final class AuthFlow: Flow {
     return self.rootViewController
   }
   
-  private lazy var rootViewController: UINavigationController = {
-    let viewController = UINavigationController()
+  private lazy var rootViewController: BaseNavigationController = {
+    let viewController = BaseNavigationController()
     return viewController
   }()
