+++ Favor/Favor/Sources/Flows/SearchFlow.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 
+import FavorUIKit
 import RxCocoa
 import RxFlow
 import RxSwift
@@ -15,8 +16,8 @@ final class SearchFlow: Flow {
   
   var root: Presentable { self.rootViewController }
   
-  private lazy var rootViewController: UINavigationController = {
-    let navigationController = UINavigationController()
+  private lazy var rootViewController: BaseNavigationController = {
+    let navigationController = BaseNavigationController()
     navigationController.isNavigationBarHidden = true
     return navigationController
   }()
